ajout rafraichissement auto map

+++ Api/ObservationModel.cs
@@ -18,9 +18,9 @@ namespace Api
 
         public string Id { get; set; }
 
-        public decimal Latitude { get; set; }
+        public double Latitude { get; set; }
 
-        public decimal Longitude { get; set; }
+        public double Longitude { get; set; }
 
         public string SpeciesName { get; set; }
